Add NewTabPagePrivacyStatsClient
